Bring back load latest button on desktop & unify impl
